Update logic to render sprites

+++ src/constants/sprites.rs
@@ -1,4 +1,4 @@
 pub const SPRITE_NUMBER: u8 = 40;
-pub const SPRITE_SIZE_BYTES: u8 = 4;
+pub const SPRITE_SIZE_BYTES: u16 = 4;
 pub const SPRITE_X_OFFSET: i32 = 8;
 pub const SPRITE_Y_OFFSET: i32 = 16;
